+++ src/main/java/com/project/app_dog_back/domain/service/IRazaService.java
@@ -1,6 +1,6 @@
 package com.project.app_dog_back.domain.service;
 
-import com.project.app_dog_back.application.dto.RazaDto;
+import com.project.app_dog_back.application.dto.catalog.RazaDto;
 import com.project.app_dog_back.domain.model.catalog.Raza;
 
 import java.util.List;
