+++ src/main/java/com/project/app_dog_back/insfraestructure/controller/catalog/RazaController.java
@@ -4,8 +4,11 @@ import com.project.app_dog_back.application.response.Meta;
 import com.project.app_dog_back.application.response.Pagination;
 import com.project.app_dog_back.application.response.Response;
 import com.project.app_dog_back.domain.model.component.TypesStatus;
+import com.project.app_dog_back.domain.service.IMetaService;
 import com.project.app_dog_back.domain.service.IRazaService;
 import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -22,17 +25,23 @@ import org.springframework.web.bind.annotation.RestController;
 @CrossOrigin(origins = "http://localhost:4200")
 public class RazaController {
     private IRazaService iRazaService;
+    private IMetaService iMetaService;
 
     @Autowired
-    public RazaController(IRazaService iRazaService) {
+    public RazaController(IRazaService iRazaService, IMetaService iMetaService) {
         this.iRazaService = iRazaService;
+        this.iMetaService = iMetaService;
     }
 
     @Operation(summary = "Get all dog breed records.")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "Successfully retrieved")
+    })
     @GetMapping
     public ResponseEntity<Response> getAll() {
         Response response = new Response();
-        response.setMeta(Meta.builder().build().toMetaBuilder(TypesStatus.SUCCESS.name()));
+        Meta meta = iMetaService.buildMetaBody("infor.queries", TypesStatus.SUCCESS.name());
+        response.setMeta(meta);
         response.setPagination(Pagination.builder().build().toPaginationBuilder());
         response.setData(iRazaService.getAll());
